Refactor and change folder name
